requirements.txt
flask==0.12.1
Jinja2==2.9.5
click==6.7
Werkzeug==0.12.1
itsdangerous==0.24
MarkupSafe==1.0
SQLAlchemy==1.1.8
flask_sqlalchemy==2.2
pynacl==1.1.2
cffi==1.10.0
pycparser==2.17
flask-wtf==0.14.2
WTForms==2.1
requests==2.13.0